Continuing restructuring of modules and models
