+++ openretina/models/sparse_autoencoder.py
