+++ openretina/modules/layers/__init__.py
@@ -1,5 +1,5 @@
 from openretina.modules.layers.ensemble import EnsembleModel
 from openretina.modules.layers.laplace import FlatLaplaceL23dnorm, Laplace
-from openretina.modules.layers.linear import Bias3DLayer, Scale2DLayer, Scale3DLayer
+from openretina.modules.layers.scaling import Bias3DLayer, Scale2DLayer, Scale3DLayer
 
 __all__ = ["FlatLaplaceL23dnorm", "Laplace", "Bias3DLayer", "Scale2DLayer", "Scale3DLayer", "EnsembleModel"]
