+++ openretina/modules/readout/factorised_gaussian.py
