+++ openretina/utils/model_utils.py
